+++ greenfield/eldenring/data.py
@@ -5531,4 +5531,4 @@ GESTURE_AWARD_FLAGS = {
     60864: (115, 1075750828, 'O Mother'),
 }
 
-_GEN_STAMP = {'inputs_hash': 'sha256:b7702a1c400a417d4c2effdbdc2df3d992d0acbe5594eeead32287ebf0a553ee', 'module': 'data.py', 'body_sha256': 'sha256:1d295bcff491ab75fab45739a053cba79b5f8218545f9adea47802a71e1c81ea'}
+_GEN_STAMP = {'inputs_hash': 'sha256:9fbc427d83b530f35af5969a4e586d7109a7669394331ab2212ebd53165024d6', 'module': 'data.py', 'body_sha256': 'sha256:1d295bcff491ab75fab45739a053cba79b5f8218545f9adea47802a71e1c81ea'}